+++ src/components/DXEEnvelopeEditor.tsx
@@ -1,35 +1,51 @@
 import { Title, Stack, Group, Text } from '@mantine/core';
 import { useState } from "react";
-import { egTypeOffsets, type egType, type VoiceParamData } from "../midi/VoiceParamData";
+//import { egTypeOffsets, type egType, type VoiceParamData } from "../midi/VoiceParamData";
+import { VoiceEditor, VoiceData, isRateParam } from '../midi/VoiceEditorData.ts';
+import type { opNumber, egParam, egType } from '../midi/VoiceEditorData.ts'; 
+
 import DXEEnvelopeGraph from "./DXEEnvelopeGraph";
 import DXEKnob from './DXEKnob';
 
 export default function DXEEnvelopeEditor(props: {
   title: string,
-  data: VoiceParamData,
+  data: VoiceData,
+  editor: VoiceEditor,
   eg: egType,
-  isTimeMode: boolean,
-  onValueChanged: (offset: number, value: number, isChangeEnd: boolean) => void })
+  isTimeMode: boolean
+  //onValueChanged: (offset: number, value: number, isChangeEnd: boolean) => void })
+  })
 {
   // Highlighted envelope parameter (0-7)
   const [highlight, setHighlight] = useState<number|undefined>(undefined);
 
-  const getVal = (offset: number) => {
-    const rawValue = props.data.getValueByOffset(
-      egTypeOffsets[props.eg] + offset);
+  const getVal = (param: egParam) => {
+    const rawValue =
+      //props.data.getValueByOffset(
+      //  egTypeOffsets[props.eg] + offset);
+      props.data.getEgValue(props.eg, param);
+
     if (props.isTimeMode) {
-      return offset < 4 ? 99-rawValue : rawValue;
+      //return offset < 4 ? 99-rawValue : rawValue;
+      return isRateParam(param) ? 99-rawValue : rawValue;
     } else {
       return rawValue
     }
   }
     
-  const setVal = (offset: number, val: number, isChangeEnd: boolean) => {
-    const fullOffset = egTypeOffsets[props.eg]+offset;
+  //const setVal = (offset: number, val: number, isChangeEnd: boolean) => {
+  const setVal = (param: egParam, val: number, isChangeEnd: boolean) => {
+    //const fullOffset = egTypeOffsets[props.eg]+offset;
+    //if (props.isTimeMode) {
+    //  props.onValueChanged(fullOffset, offset < 4 ? 99-val : val, isChangeEnd);
+    //} else {
+    //  props.onValueChanged(fullOffset, val, isChangeEnd);
+    //}
     if (props.isTimeMode) {
-      props.onValueChanged(fullOffset, offset < 4 ? 99-val : val, isChangeEnd);
-    } else {
-      props.onValueChanged(fullOffset, val, isChangeEnd);
+      props.editor.setEgValue(props.eg, param, isRateParam(param) ? 99-val : val, isChangeEnd);
+    }
+    else {
+
     }
   }
 
@@ -46,33 +62,33 @@ export default function DXEEnvelopeEditor(props: {
 
       <Group gap="sm">
         <Text className="envParamLabel col1-sm">{props.isTimeMode ? "Time" : "Rate"}</Text>
-        <DXEKnob value={getVal(0)} min={0} max={99}
-          onValueChanged={(val, ice) => setVal(0, val, ice)}
+        <DXEKnob value={getVal('R1')} min={0} max={99}
+          onValueChanged={(val, ice) => setVal('R1', val, ice)}
           onHoverChanged={(hover) => { handleHoverChanged(hover, 0); } } />
-        <DXEKnob value={getVal(1)} min={0} max={99}
-          onValueChanged={(val, ice) => setVal(1, val, ice)}
+        <DXEKnob value={getVal('R2')} min={0} max={99}
+          onValueChanged={(val, ice) => setVal('R2', val, ice)}
           onHoverChanged={(hover) => { handleHoverChanged(hover, 1); } } />
-        <DXEKnob value={getVal(2)} min={0} max={99}
-          onValueChanged={(val, ice) => setVal(2, val, ice)}
+        <DXEKnob value={getVal('R3')} min={0} max={99}
+          onValueChanged={(val, ice) => setVal('R3', val, ice)}
           onHoverChanged={(hover) => { handleHoverChanged(hover, 2); } } />
-        <DXEKnob value={getVal(3)} min={0} max={99}
-          onValueChanged={(val, ice) => setVal(3, val, ice)}
+        <DXEKnob value={getVal('R4')} min={0} max={99}
+          onValueChanged={(val, ice) => setVal('R4', val, ice)}
           onHoverChanged={(hover) => { handleHoverChanged(hover, 3); } } />
       </Group>
 
       <Group gap="sm">
         <Text className="envParamLabel col1-sm">Level</Text>
-        <DXEKnob value={getVal(4)} min={0} max={99}
-          onValueChanged={(val, ice) => setVal(4, val, ice)}
+        <DXEKnob value={getVal('L1')} min={0} max={99}
+          onValueChanged={(val, ice) => setVal('L1', val, ice)}
           onHoverChanged={(hover) => { handleHoverChanged(hover, 4); } } />
-        <DXEKnob value={getVal(5)} min={0} max={99}
-          onValueChanged={(val, ice) => setVal(5, val, ice)}
+        <DXEKnob value={getVal('L2')} min={0} max={99}
+          onValueChanged={(val, ice) => setVal('L2', val, ice)}
           onHoverChanged={(hover) => { handleHoverChanged(hover, 5); } } />
-        <DXEKnob value={getVal(6)} min={0} max={99}
-          onValueChanged={(val, ice) => setVal(6, val, ice)}
+        <DXEKnob value={getVal('L3')} min={0} max={99}
+          onValueChanged={(val, ice) => setVal('L3', val, ice)}
           onHoverChanged={(hover) => { handleHoverChanged(hover, 6); } } />
-        <DXEKnob value={getVal(7)} min={0} max={99}
-          onValueChanged={(val, ice) => setVal(7, val, ice)}
+        <DXEKnob value={getVal('L4')} min={0} max={99}
+          onValueChanged={(val, ice) => setVal('L4', val, ice)}
           onHoverChanged={(hover) => { handleHoverChanged(hover, 7); } } />
       </Group>